1.0.0

+++ PPTV/PPTV/Bussiness/Setting/SHSettingViewController.m
@@ -37,7 +37,7 @@
     
     return 10;
 }
-- (int) numberOfSectionsInTableView:(UITableView *)tableView
+- (NSInteger) numberOfSectionsInTableView:(UITableView *)tableView
 {
     return 3;
 }
